splitting matrix into hpp/cpp
